PR1741: Make PulseAudio code compliant to the JNI specification.
2011-10-03  Xerxes Ranby  <xerxes@zafena.se>
	    Robert Lougher <rob@jamvm.org.uk>

	* pulseaudio/src/native/org_classpath_icedtea_pulseaudio_Stream.c
	  (Java_org_classpath_icedtea_pulseaudio_Stream_native_1pa_1stream_1get_1buffer_1attr):
	  Make code compliant to the JNI specification.
	  Enable pulseaudio to work in combination with JVM that strictly
	  implement JNI spec.

The code is not compliant to the JNI specification.
FindClass takes a fully-qualified
classname :

http://download.oracle.com/javase/6/docs/technotes/guides/jni/spec/functions.html

name: a fully-qualified class name (that is, a package name, delimited
by “/”, followed by the class name). If the name begins with “[“ (the
array signature character), it returns an array class. The string is
encoded in modified UTF-8.

The code above is giving a type signature (qualified name inside 'L'
and ';').  HotSpot is obviously allowing this.  But this not correct
according to the specification.

Rob.